+++ tests/test_invoice.py
@@ -8,6 +8,7 @@ from django.contrib.auth import get_user_model
 from django.test.testcases import TestCase
 from stripe.error import InvalidRequestError
 
+from djstripe.enums import InvoiceStatus
 from djstripe.models import Invoice, Plan, Subscription, UpcomingInvoice
 from djstripe.settings import STRIPE_SECRET_KEY
 
@@ -380,7 +381,8 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
         default_account_mock.return_value = self.account
 
         fake_invoice = deepcopy(FAKE_INVOICE)
-        fake_invoice.update({"paid": False, "closed": False})
+        fake_invoice.update({"paid": False, "status": "open"})
+        fake_invoice.update({"auto_advance": True})
         invoice_retrieve_mock.return_value = fake_invoice
 
         invoice = Invoice.sync_from_stripe_data(fake_invoice)
@@ -474,7 +476,7 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
     @patch(
         "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
     )
-    def test_status_paid(
+    def test_status_draft(
         self,
         product_retrieve_mock,
         paymentmethod_card_retrieve_mock,
@@ -486,12 +488,11 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
     ):
         default_account_mock.return_value = self.account
 
-        invoice = Invoice.sync_from_stripe_data(deepcopy(FAKE_INVOICE))
+        invoice_data = deepcopy(FAKE_INVOICE)
+        invoice_data.update({"paid": False, "status": "draft"})
+        invoice = Invoice.sync_from_stripe_data(invoice_data)
 
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_PAID, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_PAID, invoice.legacy_status)
+        self.assertEqual(InvoiceStatus.draft, invoice.status)
 
         self.assert_fks(invoice, expected_blank_fks=self.default_expected_blank_fks)
 
@@ -536,13 +537,10 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
         default_account_mock.return_value = self.account
 
         invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False, "closed": False})
+        invoice_data.update({"paid": False, "status": "open"})
         invoice = Invoice.sync_from_stripe_data(invoice_data)
 
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_OPEN, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_OPEN, invoice.legacy_status)
+        self.assertEqual(InvoiceStatus.open, invoice.status)
 
         self.assert_fks(invoice, expected_blank_fks=self.default_expected_blank_fks)
 
@@ -574,7 +572,7 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
     @patch(
         "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
     )
-    def test_status_forgiven(
+    def test_status_paid(
         self,
         product_retrieve_mock,
         paymentmethod_card_retrieve_mock,
@@ -586,14 +584,9 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
     ):
         default_account_mock.return_value = self.account
 
-        invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False, "closed": False, "forgiven": True})
-        invoice = Invoice.sync_from_stripe_data(invoice_data)
+        invoice = Invoice.sync_from_stripe_data(deepcopy(FAKE_INVOICE))
 
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_FORGIVEN, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_FORGIVEN, invoice.legacy_status)
+        self.assertEqual(InvoiceStatus.paid, invoice.status)
 
         self.assert_fks(invoice, expected_blank_fks=self.default_expected_blank_fks)
 
@@ -625,112 +618,7 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
     @patch(
         "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
     )
-    def test_status_forgiven_deprecated(
-        self,
-        product_retrieve_mock,
-        paymentmethod_card_retrieve_mock,
-        payment_intent_retrieve_mock,
-        charge_retrieve_mock,
-        subscription_retrieve_mock,
-        balance_transaction_retrieve_mock,
-        default_account_mock,
-    ):
-        # forgiven parameter deprecated in API 2018-11-08
-        # see https://stripe.com/docs/upgrades#2018-11-08
-        default_account_mock.return_value = self.account
-
-        invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False, "closed": False})
-        invoice_data.pop("forgiven", None)  # TODO remove
-        invoice_data["status"] = "uncollectible"
-        invoice = Invoice.sync_from_stripe_data(invoice_data)
-
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_FORGIVEN, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_FORGIVEN, invoice.legacy_status)
-
-    @patch(
-        "djstripe.models.Account.get_default_account",
-        autospec=IS_STATICMETHOD_AUTOSPEC_SUPPORTED,
-    )
-    @patch(
-        "stripe.BalanceTransaction.retrieve",
-        return_value=deepcopy(FAKE_BALANCE_TRANSACTION),
-        autospec=True,
-    )
-    @patch(
-        "stripe.Subscription.retrieve",
-        return_value=deepcopy(FAKE_SUBSCRIPTION),
-        autospec=True,
-    )
-    @patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
-    @patch(
-        "stripe.PaymentIntent.retrieve",
-        return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
-        autospec=True,
-    )
-    @patch(
-        "stripe.PaymentMethod.retrieve",
-        return_value=deepcopy(FAKE_CARD_AS_PAYMENT_METHOD),
-        autospec=True,
-    )
-    @patch(
-        "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
-    )
-    def test_status_forgiven_default(
-        self,
-        product_retrieve_mock,
-        paymentmethod_card_retrieve_mock,
-        payment_intent_retrieve_mock,
-        charge_retrieve_mock,
-        subscription_retrieve_mock,
-        balance_transaction_retrieve_mock,
-        default_account_mock,
-    ):
-        # forgiven parameter deprecated in API 2018-11-08
-        # see https://stripe.com/docs/upgrades#2018-11-08
-        default_account_mock.return_value = self.account
-
-        invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False, "closed": False})
-        invoice_data.pop("forgiven", None)  # TODO remove
-        invoice = Invoice.sync_from_stripe_data(invoice_data)
-
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_OPEN, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_OPEN, invoice.legacy_status)
-
-    @patch(
-        "djstripe.models.Account.get_default_account",
-        autospec=IS_STATICMETHOD_AUTOSPEC_SUPPORTED,
-    )
-    @patch(
-        "stripe.BalanceTransaction.retrieve",
-        return_value=deepcopy(FAKE_BALANCE_TRANSACTION),
-        autospec=True,
-    )
-    @patch(
-        "stripe.Subscription.retrieve",
-        return_value=deepcopy(FAKE_SUBSCRIPTION),
-        autospec=True,
-    )
-    @patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
-    @patch(
-        "stripe.PaymentIntent.retrieve",
-        return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
-        autospec=True,
-    )
-    @patch(
-        "stripe.PaymentMethod.retrieve",
-        return_value=deepcopy(FAKE_CARD_AS_PAYMENT_METHOD),
-        autospec=True,
-    )
-    @patch(
-        "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
-    )
-    def test_status_closed(
+    def test_status_uncollectible(
         self,
         product_retrieve_mock,
         paymentmethod_card_retrieve_mock,
@@ -743,13 +631,10 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
         default_account_mock.return_value = self.account
 
         invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False})
+        invoice_data.update({"paid": False, "status": "uncollectible"})
         invoice = Invoice.sync_from_stripe_data(invoice_data)
 
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_CLOSED, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_CLOSED, invoice.legacy_status)
+        self.assertEqual(InvoiceStatus.uncollectible, invoice.status)
 
         self.assert_fks(invoice, expected_blank_fks=self.default_expected_blank_fks)
 
@@ -781,7 +666,7 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
     @patch(
         "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
     )
-    def test_status_closed_deprecated(
+    def test_status_void(
         self,
         product_retrieve_mock,
         paymentmethod_card_retrieve_mock,
@@ -791,77 +676,15 @@ class InvoiceTest(AssertStripeFksMixin, TestCase):
         balance_transaction_retrieve_mock,
         default_account_mock,
     ):
-        # closed parameter deprecated in API 2018-11-08
-        # see https://stripe.com/docs/upgrades#2018-11-08
         default_account_mock.return_value = self.account
 
         invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False})
-        invoice_data["auto_advance"] = False
-
+        invoice_data.update({"paid": False, "status": "void"})
         invoice = Invoice.sync_from_stripe_data(invoice_data)
 
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_CLOSED, invoice.status)
+        self.assertEqual(InvoiceStatus.void, invoice.status)
 
-        self.assertEqual(Invoice.STATUS_CLOSED, invoice.legacy_status)
-
-        self.assertEqual(invoice.auto_advance, invoice_data["auto_advance"])
-
-    @patch(
-        "djstripe.models.Account.get_default_account",
-        autospec=IS_STATICMETHOD_AUTOSPEC_SUPPORTED,
-    )
-    @patch(
-        "stripe.BalanceTransaction.retrieve",
-        return_value=deepcopy(FAKE_BALANCE_TRANSACTION),
-        autospec=True,
-    )
-    @patch(
-        "stripe.Subscription.retrieve",
-        return_value=deepcopy(FAKE_SUBSCRIPTION),
-        autospec=True,
-    )
-    @patch("stripe.Charge.retrieve", return_value=deepcopy(FAKE_CHARGE), autospec=True)
-    @patch(
-        "stripe.PaymentIntent.retrieve",
-        return_value=deepcopy(FAKE_PAYMENT_INTENT_I),
-        autospec=True,
-    )
-    @patch(
-        "stripe.PaymentMethod.retrieve",
-        return_value=deepcopy(FAKE_CARD_AS_PAYMENT_METHOD),
-        autospec=True,
-    )
-    @patch(
-        "stripe.Product.retrieve", return_value=deepcopy(FAKE_PRODUCT), autospec=True
-    )
-    def test_status_closed_default(
-        self,
-        product_retrieve_mock,
-        paymentmethod_card_retrieve_mock,
-        payment_intent_retrieve_mock,
-        charge_retrieve_mock,
-        subscription_retrieve_mock,
-        balance_transaction_retrieve_mock,
-        default_account_mock,
-    ):
-        # closed parameter deprecated in API 2018-11-08
-        # see https://stripe.com/docs/upgrades#2018-11-08
-        default_account_mock.return_value = self.account
-
-        invoice_data = deepcopy(FAKE_INVOICE)
-        invoice_data.update({"paid": False})
-        invoice_data.pop("auto_advance")
-        invoice_data.pop("closed", None)  # TODO remove
-        invoice_data.pop("status")
-
-        invoice = Invoice.sync_from_stripe_data(invoice_data)
-
-        with self.assertWarns(DeprecationWarning):
-            self.assertEqual(Invoice.STATUS_OPEN, invoice.status)
-
-        self.assertEqual(Invoice.STATUS_OPEN, invoice.legacy_status)
+        self.assert_fks(invoice, expected_blank_fks=self.default_expected_blank_fks)
 
     @patch(
         "djstripe.models.Account.get_default_account",